backend/requirements.txt
fastapi
uvicorn
pydantic[email]
python-multipart
motor
python-jose[cryptography]
passlib[bcrypt]
python-dotenv
httpx
livekit
livekit-api
livekit-agents[silero,turn-detector,google]~=1.4
livekit-plugins-noise-cancellation~=0.2
google-generativeai
crewai

